Fix flaky spec
failing spec: `modules/boards/spec/features/action_boards/status_type_moving_board_spec.rb:112`
failing job: https://github.com/opf/openproject/actions/runs/17290621432/job/49076891042?pr=20077

It fails because it moves the wrong card to the "Open" column. It tries
to locate the card by its text, and when it decides to move it, the card
order changes (no idea why) and both cards are swapped, resulting on the
wrong card being moved. Then the spec fails as it expects the card to be
in the other column but actually it did not move.

Sometimes, the spec passes because the card order has not changed yet
when the cards are located and moved, or the card order has already
changed before the cards are located and moved. Who knows.

Fix is to wait for the lists to properly reload after a save and moving
cards, and in the case of moving cards, it waits a little 2 secs more
just in case the order changes. It may not be reliable.