Be more lenient in the fields passed to the modal
When we do focus another input field on the page, the debounced preview logic kicks in and passes the fields name to the `frmaeSrc`. Since the search field is not whitelisted, we are throwing an argument error. This seems a bit strict.
The new logic accepts all field values. The modal is closed immediately afterwards anyways and this cancels the loading.