[#66122] Initial server rendered widgets support
Provides Project Status (BETA) widget as a Reference Implementation -
this widget is currently based on the design of the Status form on the
Project Settings > General page.

In detail:

- introduces a base `WidgetController` from which widget implementations
  should inherit.
- adds `Overviews::Widgets::ProjectStatusComponent` and accompanying
  specs. Reuses existing forms.
- adds `Overviews::Widgets::StatusController`` with custom `#update`
  action, routes and accompanying permissions.
- makes the Project Status (BETA) widget available by registering in ng
  `OpenProjectGridsModule` and Rails grid modules configuration.

N.B. The Project Status (BETA) widget is EXPERIMENTAL - it will require
further visual refinement (in a follow up work package).