Allow non-admins to see/filter created_at of a project
This is consistent with what we expose via the API, where we never tried
to hide this data. It was effectively only hidden from the UI.

We couldn't come up with a reason why it was hidden from the UI back then,
but since we now needed to expose updated_at to everyone (previous commit),
it only made sense to consistently also allow created_at to be seen by everyone.