Make result order returned by get_descendants deterministic
Previously the results were not returned in any specific order,
making the result not immediately usable in flat lists of the hierarchy.

By preordering the result, it's not necessary anymore to apply later
alphabetical ordering or similar. This is preferable, because alphabetical
ordering was only a workaround that didn't take specific ordering established
by the user into account.