Add minio to dev setup to ease development of uploads against S3
during https://community.openproject.org/wp/67403

What do you want to accomplish?

This should make it easy for people to spin up a "local S3" so they
can develop the API flow for uploading files to S3. Since this flow
is more complex than the upload to a local storage, it should be
easy to try it out locally.