Fix running Cuprite tests in Docker
This ultimatively seems to be a bug in Ferrum or an incompatibility
between Ferrum and browserless. I understand that the setting defines
whether a single or multiple web socket connections are used to communicate
with the browser, but apart from that I don't understand well what the actual
issue is.