Preload entities when rendering time entries
We were previously eager-loading work packages, but
when time entries were refactored to have a polymorphic association
to "entities", eager loading was not possible anymore and the eagerload
was removed without replacement (see eb4a1d7138).

However, preloading is a viable alternative for polymorphic associations,
so we use that now instead. Docs around the "to_eager_load" and the newly
added "to_preload" method have been added to clarify the purpose of these two
methods.