Simplify scope merging
The manual call to #includes came from a time,
when the constraint was not #merge'd, but rather
enforced through a `where(id:)` call. It's confusing today,
because the #merge call makes sure to take over all kinds of
eager loading:

* includes
* eager_load
* preload

However, the code could leave the impression, that only includes
is supported.