Revert "Revert docker workflow for 16.6"
This reverts commit 533ee53d47849b11619e415654657437f500a3e9.

What probably happened was this:

A 17.0 change was merged before the creation of the 16.6 release branch.
After the release branch got created, it contained the 17.0 change, so it
was reverted there.

Now that we merge 16.6 into dev, we need to revert that revert once more,
so that the change is not reverted in 17.0.