Fix flaky spec
spec is spec/features/projects/project_custom_fields/overview_page/dialog/inputs_spec.rb:702
run is https://github.com/opf/openproject/actions/runs/19700308113/job/56434713171?pr=21163

When opening the options of the dropdown, the field is clicked. If unlucky, it can click on the "X" of an option already selected and remove it. When this happens, it does not open the dropdown (this explains why the previous run was flaky).

The default implementation in both Cuprite and Selenium is to click in
the middle of the element. I don't understand why we can't reproduce it
though.

Fix is to click the arrow directly.