Make 2+ shared work packages accessible to non-project members
https://community.openproject.org/wp/68921

The condition was excluding work packages from a project if there was
more than one work package shared within the same project, and the user
was not a member of the project.

Reworked the query into a simpler sql applying union to get set of
member or entity permitted ids.

Co-authored-by: Jens Ulferts <jens.ulferts@googlemail.com>